# api-spec.yaml
openapi: 3.0.3
info:
  title: Minecraft Server Manager API
  description: RESTful API for managing Minecraft servers, user authentication, and administrative functions
  version: 0.1.0-alpha
  contact:
    name: Minecraft Server Manager
    url: https://github.com/your-username/mcServerManager
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://your-domain.com/api/v1
    description: Production server
security:
  - sessionAuth: []
paths:
...
  /admin/config:
    get:
      tags:
        - Admin
      summary: Get system configuration
      description: Get system configuration (admin only)
      responses:
        '200':
          description: System configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    put:
      tags:
        - Admin
      summary: Update system configuration
      description: Update system configuration (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigUpdateRequest'
      responses:
        '200':
          description: Configuration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/stats:
    get:
      tags:
        - Admin
      summary: Get system statistics
      description: Get system statistics (admin only)
      responses:
        '200':
          description: System statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users:
    get:
      tags:
        - Admin
      summary: List users
      description: List all users (admin only)
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Admin
      summary: Create user
      description: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /admin/users/{id}:
    put:
      tags:
        - Admin
      summary: Update user
      description: Update user details (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Admin
      summary: Delete user
      description: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session-based authentication
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error description
        error:
          type: string
          example: Detailed error information

    # User schemas
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        is_admin:
          type: boolean
          example: true
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-09-04T16:20:00.000Z"
        last_login:
          type: string
          format: date-time
          example: "2025-09-04T16:20:00.000Z"
    UserResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
    UserCreateRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: newuser
        password:
          type: string
          format: password
        email:
          type: string
          format: email
          example: user@example.com
        is_admin:
          type: boolean
          example: false
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          example: updateduser
        email:
          type: string
          format: email
          example: updated@example.com
        is_admin:
          type: boolean
          example: false
        is_active:
          type: boolean
          example: true
    UsersResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            users:
              type: array
              items:
                $ref: '#/components/schemas/User'

    # Authentication schemas
...
    Config:
      type: object
      properties:
        app_title:
          type: string
          example: Minecraft Server Manager
        server_hostname:
          type: string
          example: localhost
        max_total_mb:
          type: integer
          example: 8192
        max_server_mb:
          type: integer
          example: 4096
        default_server_mb:
          type: integer
          example: 1024
    ConfigResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            config:
              $ref: '#/components/schemas/Config'
    ConfigUpdateRequest:
      type: object
      properties:
        app_title:
          type: string
          example: My Server Manager
        server_hostname:
          type: string
          example: myserver.com
        max_total_mb:
          type: integer
          example: 16384
        max_server_mb:
          type: integer
          example: 8192
    StatsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            stats:
              type: object
              properties:
                total_servers:
                  type: integer
                  example: 5
                running_servers:
                  type: integer
                  example: 2
                total_users:
                  type: integer
                  example: 3
                total_memory_allocated:
                  type: integer
                  example: 3072
                system_memory_usage:
                  type: number
                  example: 45.5
                disk_usage:
                  type: object
                  properties:
                    total_gb:
                      type: integer
                      example: 100
                    used_gb:
                      type: integer
                      example: 25
                    free_gb:
                      type: integer
                      example: 75
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...